dom/base/test/test_bug425013.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=425013
-->
<head>
  <title>Test for Bug 425013</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="runtests();">

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=425013">Mozilla Bug 425013</a>
<br>

<script type="text/javascript;version=1.7">
var missingPlugins = new Array();
var OBJLC = SpecialPowers.Ci.nsIObjectLoadingContent;

function pluginBinding(event)
{
  var plugin = event.target;
  plugin instanceof OBJLC;
  if (SpecialPowers.wrap(plugin).pluginFallbackType == OBJLC.PLUGIN_UNSUPPORTED)
    missingPlugins.push(plugin);
}

document.addEventListener("PluginBindingAttached", pluginBinding, true);
</script>

<object type="foo/bar" id="obj1"></object>

<object type="foo/bar" id="obj2">
  <embed type="a/b" id="embed1"></embed>
</object>

<object type="foo/bar"
        data="data:foo/bar,bah" id="obj3">
<param name="movie" value="data:foo/bar,bah">
</object>

<object type="foo/bar"
        data="data:foo/bar,bah" id="obj4">
<param name="movie" value="data:foo/bar,bah">
<p>Alternate content</p>
</object>

<object type="text/html"
        data="data:text/html,an html document in an object tag" id="obj5">
</object>

<object type="text/html"
        data="data:text/html,an html document in an object tag" id="obj6">
<p>Alternate content</p>
</object>

<embed type="a/b" id="embed2"></embed>

<script class="testbody" type="text/javascript">
function runtests()
{
  // Force layout flush so the binding is attached and the event is fired
  document.getElementById("obj1").clientTop;
  document.getElementById("obj3").clientTop;
  document.getElementById("embed1").clientTop;
  document.getElementById("embed2").clientTop;

  // Let pending events flush
  SimpleTest.executeSoon(function () {
    ok(missingPlugins.indexOf(document.getElementById("obj1")) >= 0, "Missing plugin element obj1");
    ok(missingPlugins.indexOf(document.getElementById("embed1")) >= 0, "Missing plugin element embed1");
    ok(missingPlugins.indexOf(document.getElementById("embed2")) >= 0, "Missing plugin element embed2");
    ok(missingPlugins.indexOf(document.getElementById("obj3")) >= 0, "Missing plugin element obj3");

    is(missingPlugins.length, 4, "Wrong number of missing plugins");

    SimpleTest.finish();
  });
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
